scripts/dataPromotion/loadClinical/requirements.txt
python-dotenv==0.15.0
mysql_connector_repackaged==0.3.1
mysql-connector-python
